completed issue #1 separate the header
