+++ client/App.js
@@ -1,45 +1,38 @@
 import React from 'react';
 
-import { StyleSheet, SafeAreaView, AppRegistry } from 'react-native';
+import { AppRegistry } from 'react-native';
 import { DefaultTheme, Provider as PaperProvider } from 'react-native-paper';
 import { expo } from './app.json';
 
-import Main from './src/page/main';
+import '@expo/match-media'
+import { useMediaQuery } from "react-responsive";
+
+import MainMobile from './src/mobile/main';
+import MainWeb from './src/web/main';
+import { paperTheme } from './src/styles/app';
 
 export default function App() {
-  // COMMENT OUT ALL console.log statements IN PRODUCTION
-  console.log("App executed");
-  let x = 1;          // testing: try setting a break point here
+  // note: this must go inside the body of a function component. Top level.
+  const isTabletOrMobileDevice = useMediaQuery({    
+    maxDeviceWidth: 1224,
+    // alternatively...
+    query: "(max-device-width: 1224px)"  
+  });
 
-  // testing: create a bug
-    // let x;
-    // x.toString();
+  function Body() {
+    if (isTabletOrMobileDevice) {
+      // note: don't use react-router-native bc it will crash the web build
+      return <MainMobile/>
+    }
+    // use react-router (for web) for page like navigation
+    return <MainWeb/>
+  }
 
   return (
-    <PaperProvider theme={theme}>
-        <Main/>
+    <PaperProvider theme={paperTheme}>
+      <Body/>
     </PaperProvider>
   );
 }
 
 AppRegistry.registerComponent(expo.name, () => App);
-
-// This theme applies to react-native-paper components only
-const theme = {
-  ...DefaultTheme,
-  roundness: 2,
-  colors: {
-    ...DefaultTheme.colors,
-    primary: 'tomato',
-    accent: 'yellow',  
-  },
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-})
