+++ client/package.json
@@ -17,8 +17,7 @@
     "react-native-paper": "^4.7.2",
     "react-native-web": "^0.15.6",
     "react-responsive": "^8.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-router-native": "^5.2.0"
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0"
